+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:39441/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:40293/jest?","mongoDBName":"jest"}
