+++ src/main/factories/signup/signup-validation.spec.ts
@@ -1,9 +1,11 @@
-import { CompareFieldValidation } from '../../../presentation/helpers/validators/compare-fields-validation'
-import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '../../../presentation/helpers/validators/required-field-validation'
-import { Validation } from '../../../presentation/protocols/validation'
-import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
+import {
+  CompareFieldValidation,
+  EmailValidation,
+  RequiredFieldValidation,
+  ValidationComposite,
+} from '../../../presentation/helpers/validators'
 import { EmailValidator } from '../../../presentation/protocols'
+import { Validation } from '../../../presentation/protocols/validation'
 import { makeSignUpValidation } from './signup-validation'
 
 jest.mock('../../../presentation/helpers/validators/validation-composite')
