+++ src/main/factories/signup/signup-validation.ts
@@ -1,8 +1,10 @@
-import { CompareFieldValidation } from '../../../presentation/helpers/validators/compare-fields-validation'
-import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '../../../presentation/helpers/validators/required-field-validation'
+import {
+  CompareFieldValidation,
+  EmailValidation,
+  RequiredFieldValidation,
+  ValidationComposite,
+} from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
 import { EmailValidatorAdapter } from '../../../utils/email-validator.adapter'
 
 export const makeSignUpValidation = (): ValidationComposite => {
